feat(alerts): implement AlertRepository for managing alerts and update AlertsPage to use repository data
