+++ lib/screens/home/home_screen.dart
@@ -6,6 +6,7 @@ import 'alerts_page.dart';
 import 'estoque_categorias_page.dart';
 import '../../services/auth_service.dart';
 import '../login/login_screen.dart';
+import '../../repositories/alert_repository.dart';
 
 // Classe para desenhar o gráfico de pizza do estoque
 class PieChartPainter extends CustomPainter {
@@ -117,10 +118,18 @@ class _HomeScreenState extends State<HomeScreen>
       vsync: this,
       duration: const Duration(milliseconds: 300),
     );
+    // Escutar mudanças na contagem de alertas para atualizar o dashboard
+    AlertRepository.instance.countNotifier.addListener(_onAlertsCountChanged);
+  }
+
+  void _onAlertsCountChanged() {
+    if (mounted) setState(() {});
   }
 
   @override
   void dispose() {
+    // Remover listener para evitar leaks
+    AlertRepository.instance.countNotifier.removeListener(_onAlertsCountChanged);
     _animationController.dispose();
     super.dispose();
   }
@@ -453,7 +462,7 @@ class _HomeScreenState extends State<HomeScreen>
               color2: Colors.green.shade200,
             ),
 
-            // Card de Alertas (estoque baixo / vencimento)
+            // Card de Alertas (estoque baixo / vencimento) - usa contagem do repositório
             _buildDashboardCard(
               'Alertas (estoque baixo / vencimento)',
               Icons.warning_amber,
@@ -464,7 +473,8 @@ class _HomeScreenState extends State<HomeScreen>
                   MaterialPageRoute(builder: (_) => const AlertsPage()),
                 );
               },
-              hasAlert: true,
+              hasAlert: AlertRepository.instance.countNotifier.value > 0,
+              alertCount: AlertRepository.instance.countNotifier.value,
               color2: Colors.red.shade200,
             ),
           ],
@@ -708,8 +718,7 @@ class _HomeScreenState extends State<HomeScreen>
             ),
           ),
         ),
-      ),
-    );
+    ));
   }
 
   // Widget para mostrar uma estatística no card de estoque
@@ -760,6 +769,7 @@ class _HomeScreenState extends State<HomeScreen>
     VoidCallback onTap, {
     bool hasAlert = false,
     Color? color2,
+    int? alertCount, // novo parâmetro opcional
   }) {
     final gradientColor = color2 ?? color.withOpacity(0.6);
 
@@ -839,9 +849,9 @@ class _HomeScreenState extends State<HomeScreen>
                             ),
                           ],
                         ),
-                        child: const Text(
-                          '3',
-                          style: TextStyle(
+                        child: Text(
+                          '${alertCount ?? AlertRepository.instance.countNotifier.value}',
+                          style: const TextStyle(
                             color: Colors.white,
                             fontWeight: FontWeight.bold,
                             fontSize: 12,
@@ -872,8 +882,7 @@ class _HomeScreenState extends State<HomeScreen>
             ),
           ),
         ),
-      ),
-    );
+    ));
   }
 }
